ethdev: fix getting supported ptypes list
Incorrect ptypes list causes buffer overflow for Address Sanitizer
run.Previously, the last element in the ptypes lists to be
"RTE_PTYPE_UNKNOWN" for rte_eth_dev_get_supported_ptypes(), but this was
not clearly documented and many PMDs did not follow this implementation.
Instead, the dev_supported_ptypes_get() function pointer now returns the
number of elements to eliminate the need for "RTE_PTYPE_UNKNOWN"
as the last item.

Fixes: 4790935 ("ethdev: make device operations struct private")
Cc: ferruh.yigit@intel.com
CC: stable@dpdk.org
Signed-off-by: 0-day Robot <robot@bytheb.org>