net/iavf: fix vf startup coredump
When the vf starts to request multiple queues, the pf sends a reset
command to the vf. During the reset process, adminq sends an abnormal
message to pf for an unknown reason, and the resource request fails
resulting in a coredump.

This patch fixes the issue by checking the reset state before resetting.

v2: fix windows compilation errors

Signed-off-by: Shiyang He <shiyangx.he@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>