net/mlx5: fix secondary process query stats segfault
The "outer_of_buffer" counter is owned by the primary process devx
object and it is pointer by pointer in mlx5_priv structure. Actually,
there are 4 levels' pointers in this piece of code.

The secondary process can't access this part directly since it belongs
to another process' heap.

Return ENOTSUP as workaround.

Signed-off-by: Rongwei Liu <rongweil@nvidia.com>
Acked-by: Matan Azrad <matan@nvidia.com>
Fixes: 750e48c7d ("common/mlx5: add DevX commands for queue counters")
Cc: matan@nvidia.com
Cc: stable@dpdk.org
Signed-off-by: 0-day Robot <robot@bytheb.org>