net/ixgbe: increase vf reset timeout
When vf issues a reset to pf there is a 50 msec wait plus an additional
max of 1 msec for the pf to indicate the reset is complete before
timeout.

In some cases, it is seen that the reset is timing out, in which case
the reset does not complete and an error is returned.

In order to account for this, continue to wait an initial 50 msecs,
but then allow a max of an additional 50 msecs for the command to
complete.

Fixes: af75078fece3 ("first public release")

Signed-off-by: Kevin Traynor <ktraynor@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>