net/hns3: fix Rx packet truncation when KEEP CRC enabled
When KEEP_CRC offload is enabled, some packets will be truncated and
the CRC is still be stripped in following cases:
1. For HIP08 hardware, the packet type is TCP and the length
   is less than or equal to 60B.
2. For other hardwares, the packet type is IP and the length
   is less than or equal to 60B.

So driver has to recaculate positively the CRC of these packets
on above cases.

In addition, to avoid impacting performance, KEEP_CRC is not
supported when NEON or SVE algorithm is used.

Fixes: 8973d7c4ca12 ("net/hns3: support keeping CRC")
Cc: stable@dpdk.org

Signed-off-by: Dengdui Huang <huangdengdui@huawei.com>
Signed-off-by: Jie Hai <haijie1@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>